requirements.txt
discord.py~=2.7.1
python-dotenv~=1.2.2
Flask~=3.1.3